requirements.txt
Django==2.2.3
djangorestframework
django-crispy-forms==1.7.2
gunicorn==19.9.0
